better logging, vecs are weird

+++ src/headers/libs/libstd/itoa.hpp
@@ -1,6 +1,8 @@
 #ifndef LIBS_LIBSTD_ITOA_HPP
 #define LIBS_LIBSTD_ITOA_HPP
 
-int itoa(int value, char *sp, int radix=10);
+#include <stdint.h>
+
+int itoa(int64_t value, char *sp, int radix=10);
 
 #endif // !LIBS_LIBSTD_ITOA_HPP
